+++ Components/ts/Once/BrowserKernel.ts
