--- EAMD.ucp/BaseKernel.js
@@ -1,9 +1,0 @@
-/* eslint-disable no-unused-vars */
-/* eslint-disable no-var */
-export class BaseKernel {
-    global;
-    constructor(gt) {
-        this.global = gt;
-        this.global.ONCE = this;
-    }
-}
