--- EAMD.ucp/BrowserKernel.js
@@ -1,10 +1,0 @@
-import { BaseKernel } from './BaseKernel.js';
-export class BrowserKernel extends BaseKernel {
-    constructor() {
-        super(window);
-    }
-    async start() {
-        console.log('running in browser');
-        return this;
-    }
-}
