--- EAMD.ucp/NodeJsKernel.js
@@ -1,6 +1,0 @@
-import { BaseKernel } from './BaseKernel.js';
-export class NodeJsKernel extends BaseKernel {
-    constructor() {
-        super(global);
-    }
-}
