+++ package.json
@@ -27,12 +27,15 @@
   },
   "homepage": "https://github.com/Cerulean-Circle-GmbH/once.ts#readme",
   "devDependencies": {
+    "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/node": "^17.0.5",
     "@types/node-forge": "^0.10.10",
     "@types/pem": "^1.9.6",
+    "@types/serve-index": "^1.9.1",
     "@typescript-eslint/eslint-plugin": "^5.8.1",
     "@typescript-eslint/parser": "^5.8.1",
+    "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "eslint": "^7.32.0",
     "eslint-config-standard": "^16.0.3",
@@ -44,6 +47,7 @@
     "nodemon": "^2.0.15",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
+    "serve-index": "^1.9.1",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.4"
   }
